[REF] *: Remove UIActions from core plugins
Core plugins should no longer interact with the UI, so we removed
UIActions from core plugins. They are now accessible by UI plugins
only.
